Add files via upload
In main_final, there is a line of code that might cause an error (this is due to one step in the code, which fails when run repeatedly). This version runs smoothly to the regression results the first time, but during the second run, the intermediate process (the construction of turnover factors) may produce an error. It needs to be run after using clear.